+++ components/Controllers/OpenVault/VaultDebtSlider/VaultDebtSlider.tsx
@@ -1,10 +1,8 @@
 import useResizeObserver from '@react-hook/resize-observer';
 import { Slider } from 'components/Slider';
-import { Tooltip, TooltipReference as TTR } from 'components/Tooltip';
 import { useTheme } from 'hooks/useTheme';
 import { formatPercent } from 'lib/numberFormat';
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { useTooltipState } from 'reakit/Tooltip';
 
 import styles from '../VaultDebtPicker/VaultDebtPicker.module.css';
 
@@ -149,44 +147,5 @@ export function VaultDebtSlider({
         theme={theme}
       />
     </div>
-  );
-}
-
-type VaultDebtExplainerProps = {
-  daysToLiquidation: number;
-  liquidationTriggerPrice: string;
-  maxLTV: number;
-  currentLTV: number;
-};
-function VaultDebtExplainer({
-  currentLTV,
-  daysToLiquidation,
-  maxLTV,
-  liquidationTriggerPrice,
-}: VaultDebtExplainerProps) {
-  const currentLTVTooltip = useTooltipState();
-  const accruingInterestTooltip = useTooltipState();
-  const nftValueTooltip = useTooltipState();
-  return (
-    <>
-      <p>
-        Loan liquidates when <TTR {...currentLTVTooltip}>Current LTV</TTR> (
-        {formatPercent(currentLTV)}) reaches Max LTV ({formatPercent(maxLTV)}).
-        This can happen by{' '}
-        <TTR {...accruingInterestTooltip}>accruing interest</TTR> or by a drop
-        in <TTR {...nftValueTooltip}>NFT value</TTR>.
-      </p>
-      <Tooltip {...currentLTVTooltip}>TODO</Tooltip>
-      <Tooltip {...accruingInterestTooltip}>
-        {daysToLiquidation === 0
-          ? 'The current interest rate is negative, and is not currently moving this loan closer to liquidation.'
-          : `Assuming today's interest rate and NFT price, interest charges will result in liquidation after ${daysToLiquidation} days.`}
-      </Tooltip>
-      <Tooltip {...nftValueTooltip}>
-        If contract&apos;s valuation of the collateral drops below{' '}
-        {liquidationTriggerPrice}, this loan will be liquidated at auction to
-        cover the debt.
-      </Tooltip>
-    </>
   );
 }
